Change default port and add some basic testings
Make it possible for the site to run without extra configuration when
deployed to production, so extra configuration environment variables
need to be set only on development machine.

Also change the 'PORT' variable name, if this ever runs under Heroku,
it needs to be called 'PORT' as other name doesn't work.

New test files went down with this commit and I'm too lazy to revert
now, so this commit adds also Mocha and Chai libraries and a couple of
page tests which are not connected yet.